Update main.html
Added 10 lines of code to the bottom of the html file. I chose to place it here in case anyone wants to move it around, since I didn't want to mess anything up. Most of these lines are just referencing to the other features that need to be linked to it